Fix bugs in index and delete
- Looked into a problem on index where the reverse would flip back and
forth -> moved the .reverse() down to the return rather than before
the map, this works
- Looked at the problem with delete where it would Redirect the user to
the index; however, the previously deleted post would still be
there. -> commented out the Redirect call at the bottom and rather,
deconstructed history from this.props and then used history.push() to
send the user back to the index after the delete has occurred.

co-authored-by: Pete V. petevallerie@gmail.com
co-authored-by: Faith F. faithfitts@gmail.com
co-authored-by: Nicole S. nicolesurawski@gmail.com